replace FireBase
