+++ src/routes/GamePage/index.js
@@ -3,21 +3,7 @@ import s from './style.module.css'
 import POKEMONS from "../../data/POKEMONS.json";
 import PokemonCard from "../../components/PokemonCard";
 import {useEffect, useState} from 'react'
-
-import firebase from "firebase";
-
-const firebaseConfig = {
-    apiKey: "AIzaSyBmYiwrKAK4VuMaxmQXZzD5CYdUOLx6078",
-    authDomain: "pokemon-cardz.firebaseapp.com",
-    databaseURL: "https://pokemon-cardz-default-rtdb.firebaseio.com",
-    projectId: "pokemon-cardz",
-    storageBucket: "pokemon-cardz.appspot.com",
-    messagingSenderId: "258144951272",
-    appId: "1:258144951272:web:53b2060a3ca2ba60b2625d"
-};
-
-firebase.initializeApp(firebaseConfig);
-const database = firebase.database();
+import database from "../../service/firebase";
 
 database.ref('pokemons').once('value', (snapshot) => {
     console.log('####DATABase:', snapshot.val())
